fix(memory): close Round 1 review findings for PR #1638
Apply Round 1 closure for swarm-pr-review at
.swarm/pr-review/pr-1638-202607021700/feedback-handoff.md.

Resolves 5 confirmed findings from the review; 1 LOW finding
(branch is 62 commits behind) deferred to author; ~30 hermes
CRITICAL/HIGH findings DISPROVED by my independent validation.

FB-001 [CRITICAL] — MIGRATION version-number collision with PR #1636
- src/memory/sqlite-provider.ts: renumbered MIGRATIONS v7→v9, v8→v10 to
  avoid collision with already-merged PR #1636. SQL unchanged. The
  `runMigrations` skip-when-version-≤-currentVersion logic no longer
  drops PR #1638's `memory_reward_events` table and `unit_id` column on
  databases initialized from `origin/main`.
- tests/unit/memory/migration-renumber.test.ts: NEW — 5 tests covering
  v9/v10 migration application, table creation, column addition, end-to-end
  API usage, and idempotency. Verifies upgrade from main-initialized DB
  exercises `appendRewardEvent` and `listRecallUsage({unitId})`.

FB-002 [HIGH] — Test file 500-line violations
- tests/unit/memory/finalize-reward-sweep.test.ts: DELETED (was 654 lines)
- tests/unit/memory/finalize-reward-sweep-basic.test.ts: NEW (290 lines) — happy path + memory-disabled + non-blocking
- tests/unit/memory/finalize-reward-sweep-edge-cases.test.ts: NEW (243 lines) — SC-007 negative reward + edge cases
- tests/unit/memory/finalize-reward-sweep-multi-runid.test.ts: NEW (330 lines) — multi-runId + propagation
- tests/unit/memory/reward-capture-propagation.test.ts: DELETED (was 589 lines)
- tests/unit/memory/reward-capture-council.test.ts: NEW (276 lines) — basic capture
- tests/unit/memory/reward-capture-propagation-only.test.ts: NEW (320 lines) — soft Q-propagation
- All 7 new files under 500 lines (PR body O-007 claim "largest 435" was
  materially false; actual largest was 654 lines).

FB-003 [MEDIUM] — JSON truncation creates invalid JSON
- src/memory/reward-capture.ts: NEW `truncateObjectForJson<T>` helper that
  truncates the source object before stringifying, producing valid JSON
  even when truncation triggers. 3-strategy: unchanged / with-marker /
  placeholder.
- src/tools/submit-phase-council-verdicts.ts: applied at line 348
  (replaced `slice(0, cap) + '…[truncated]'` with
  `JSON.stringify(truncateObjectForJson(...))`).
- src/hooks/delegation-gate.ts: applied at line 1375 (same fix).
- tests/unit/tools/submit-phase-council-verdicts-reward.test.ts: added 3
  regression tests verifying JSON parseability after truncation.
- tests/unit/hooks/delegation-gate-council-reward.test.ts: added 3
  regression tests.

FB-004 [MEDIUM] — applyCouncilReward read-then-write race condition
- src/memory/provider.ts: added `MemoryTransaction` interface and
  `withTransaction` method to `MemoryProvider` (backward-compatible
  optional).
- src/memory/sqlite-provider.ts: implemented `withTransaction` using
  `db.transaction()`.
- src/memory/reward-capture.ts: wrapped `applyCouncilReward`'s
  read-then-update loop in `withTransaction` (with graceful degradation
  if provider doesn't implement it).
- tests/unit/memory/reward-capture-fb004-fb005.test.ts: NEW (313 lines)
  — regression test verifying `withTransaction` is called and qBefore
  is read correctly inside the transaction.

FB-005 [MEDIUM] — buildRetrievalRecency iterates ALL recall events unboundedly
- src/memory/provider.ts: added `since` field to `MemoryRecallUsageFilter`.
- src/memory/sqlite-provider.ts: implemented `since` WHERE clause in
  `listRecallUsage`.
- src/memory/reward-capture.ts: added `RECENCY_WINDOW_MS` (30 days, matches
  PR #1638 config default) and `RECENCY_EVENT_LIMIT` (1000) constants;
  updated `buildRetrievalRecency` to pass `since` filter + limit.
- tests/unit/memory/reward-capture-fb004-fb005.test.ts: regression tests
  for the `since` filter behavior.

PRE_EXISTING (FB-006): NOT addressed in this commit
- Branch is 62 commits behind `origin/main`. Full rebase out of scope
  (would require significant conflict resolution with main's v1-v8
  migrations). Author should rebase before merging.

CARRIED FORWARD (FB-007): DISPROVED
- ~30 hermes CRITICAL/HIGH findings were DISPROVED by my independent
  validation:
  - MemoryRewardEvent import path wrong: types ARE in provider.ts:46-62
  - MIGRATIONS missing v0-v6: v2 is intentional LEGACY_JSONL_MIGRATION_VERSION
  - runFinalizeRewardSweep ordering: function never re-throws (try/catch)
  - Q-value boost formula inverted: formula is `(qValue - NEUTRAL_Q_VALUE) * weight`
  - QLearningConfig missing defaults: DEFAULT_QLEARNING_CONFIG provides all
  - backfillRecallRunIds DB connection leak: uses this.requireDb()
  - ctx.sessionID not validated: truthy check exists
  - `_meta` table not created: created by v4 migration
  - Test files contain no test cases: files have substantive content (269-330 lines)

DEFERRED (FB-008): minor concerns from hermes
- Test docs/comment mismatches in registration-parity.test.ts
- Some EBUSY/leak risks in cross-platform scenarios
- Test infrastructure gaps
- These are minor and can be addressed in follow-up

Test timestamp updates (incidental fixes during validation)
- tests/unit/memory/finalize-reward-sweep-multi-runid.test.ts:
  updated `'2026-05-25T00:00:00.000Z'` to `'2026-06-20T00:00:00.000Z'`
  (within 30-day recency window per PR config).
- tests/unit/memory/reward-capture-council.test.ts: same update.
- tests/unit/memory/reward-capture-propagation-only.test.ts:
  updated `WITHIN_WINDOW` constant to `'2026-06-25T00:00:00.000Z'`.
- These were stale test data from PR #1638's authoring; the hardcoded
  dates were outside the documented 30-day recency window by the time
  the tests ran. Updated to dates within the window.

Validation:
- git diff --check: clean
- bun run typecheck: exit 0
- bun run drift:check: PASS (no new warnings)
- bunx biome ci: clean (no fixes applied)
- 39 tests in modified test files: ALL PASS
  (was 24 before this commit; +15 new regression tests)
- Pre-existing failures in tests/unit/memory/ confirmed reproduce on
  origin/main (15 fails acknowledged in PR body)
